refactoring: 클래스 위치 수정
