+++ src/main/java/laddarGame/domain/Ladder.java
@@ -1,4 +1,4 @@
-package laddarGame.controller;
+package laddarGame.domain;
 
 import java.util.List;
