+++ src/main/java/laddarGame/domain/Line.java
@@ -1,4 +1,4 @@
-package laddarGame.controller;
+package laddarGame.domain;
 
 import java.util.List;
